+++ examples/with-nuxtjs/mocks/users/_userId.js
@@ -1,7 +1,0 @@
-const users = [{ id: 0, name: 'foo' }]
-
-export default {
-  get({ values }) {
-    return [200, users.find(user => user.id === values.userId)]
-  }
-}