+++ jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  testPathIgnorePatterns: ['examples'],
-  moduleNameMapper: { '^~/(.+)': '<rootDir>/packages/$1' }
+  testPathIgnorePatterns: ['samples']
 }