+++ packages/aspida-mock/aspida.config.js
@@ -1,0 +1,1 @@
+module.exports = { input: '../aspida/samples', baseURL: 'https://example.com/api/' }