+++ packages/aspida-mock/src/cli/writeRouteFile.ts
@@ -1,0 +1,7 @@
+import fs from 'fs'
+import { Template } from './Template'
+
+export default ({ filePath, text }: Template) => {
+  fs.writeFileSync(filePath, text, 'utf8')
+  console.log(`${filePath} was built successfully.`)
+}