+++ packages/aspida/src/createMethodsString.ts
@@ -1,8 +1,8 @@
 import { Project } from 'ts-morph'
 import { SyntaxKind } from 'typescript'
+import { LowerHttpMethod, AspidaMethodParams } from './'
 
-type HttpMethod = 'get' | 'post' | 'put' | 'delete' | 'head' | 'patch'
-type MethodsProperties = 'query' | 'reqHeaders' | 'reqData' | 'reqType' | 'resHeaders' | 'resData'
+type MethodsProperties = keyof AspidaMethodParams
 
 export default (target: string, indent: string, importName: string, newUrl: string) => {
   const source = new Project().addSourceFileAtPath(target)
@@ -11,7 +11,7 @@ export default (target: string, indent: string, importName: string, newUrl: stri
   if (!methodsInterface) return ''
 
   methodsInterface.getProperties().forEach(method => {
-    const methodName = method.getName() as HttpMethod
+    const methodName = method.getName() as LowerHttpMethod
     const typeInfo: { [key in MethodsProperties]?: { value: string; hasQuestion: boolean } } = {}
     method.getChildrenOfKind(SyntaxKind.TypeLiteral).forEach(n =>
       n.getProperties().forEach(p => {
@@ -22,47 +22,42 @@ export default (target: string, indent: string, importName: string, newUrl: stri
       })
     )
 
-    const hasOption = typeInfo.query || typeInfo.reqData || typeInfo.reqHeaders
     const isOptionRequired =
       (typeInfo.query && !typeInfo.query.hasQuestion) ||
       (typeInfo.reqData && !typeInfo.reqData.hasQuestion) ||
       (typeInfo.reqHeaders && !typeInfo.reqHeaders.hasQuestion)
 
-    const reqData = (method: HttpMethod) =>
+    const reqData = (method: LowerHttpMethod) =>
       typeInfo.reqData
         ? ` data${typeInfo.reqData.hasQuestion ? '?' : ''}: ${importName}['${method}']['reqData'],`
         : ''
-    const query = (method: HttpMethod) =>
+    const query = (method: LowerHttpMethod) =>
       typeInfo.query
         ? ` query${typeInfo.query.hasQuestion ? '?' : ''}: ${importName}['${method}']['query'],`
         : ''
-    const reqHeaders = (method: HttpMethod) =>
+    const reqHeaders = (method: LowerHttpMethod) =>
       typeInfo.reqHeaders
         ? ` headers${
             typeInfo.reqHeaders.hasQuestion ? '?' : ''
           }: ${importName}['${method}']['reqHeaders'],`
         : ''
-    const resHeaders = (method: HttpMethod) =>
+    const resHeaders = (method: LowerHttpMethod) =>
       typeInfo.resHeaders ? `, ${importName}['${method}']['resHeaders']` : ''
-    const option = (method: HttpMethod) =>
-      hasOption
-        ? `option${isOptionRequired ? '' : '?'}: {${`${reqData(method)}${query(method)}${reqHeaders(
-            method
-          )}`.slice(0, -1)} }`
-        : ''
+    const option = (method: LowerHttpMethod) =>
+      `option${isOptionRequired ? '' : '?'}: {${reqData(method)}${query(method)}${reqHeaders(
+        method
+      )} config?: U }`
     const request = () =>
-      !hasOption
-        ? ''
-        : !typeInfo.reqData
-        ? ', option'
-        : `, ${isOptionRequired ? '' : '!option ? undefined : '}optionToRequest(option${
-            typeInfo.reqType
-              ? `, '${typeInfo.reqType.value}'`
-              : typeInfo.reqData && /^(ArrayBuffer|Blob|string)$/.test(typeInfo.reqData.value)
-              ? `, '${typeInfo.reqData.value}'`
-              : ''
-          })`
-    const resData = (method: HttpMethod) =>
+      `, option${
+        !typeInfo.reqData
+          ? ''
+          : typeInfo.reqType
+          ? `, '${typeInfo.reqType.value}'`
+          : typeInfo.reqData && /^(ArrayBuffer|Blob|string)$/.test(typeInfo.reqData.value)
+          ? `, '${typeInfo.reqData.value}'`
+          : ''
+      }`
+    const resData = (method: LowerHttpMethod) =>
       `${typeInfo.resData ? `${importName}['${method}']['resData']` : 'void'}`
     const resMethodName = () =>
       !typeInfo.resData
@@ -73,9 +68,11 @@ export default (target: string, indent: string, importName: string, newUrl: stri
 
     const tmpChanks = [
       `(${option(methodName)}) =>`,
-      `client.fetch<${resData(methodName)}${resHeaders(
-        methodName
-      )}>(\`$\{prefix}${newUrl}\`, '${methodName.toUpperCase()}'${request()}).${resMethodName()}()`
+      `client.fetch<${resData(methodName)}${resHeaders(methodName)}>(prefix, ${
+        newUrl.includes('${') ? '`' : "'"
+      }${newUrl}${
+        newUrl.includes('${') ? '`' : "'"
+      }, '${methodName.toUpperCase()}'${request()}).${resMethodName()}()`
     ]
 
     chanks.push(`${indent}  ${methodName}: ${tmpChanks[0]}