+++ packages/axios-mock-server/samples/index.ts
@@ -1,5 +1,0 @@
-import { MockMethods } from '../src/types'
-
-export default {
-  get: () => [200, 'Hello world!']
-} as MockMethods