+++ packages/axios-mock-server/src/asyncResponse.ts
@@ -1,7 +1,0 @@
-import { MockResponse } from './types'
-
-export default async (
-  status: number,
-  query: Promise<any>,
-  headers?: any
-): Promise<MockResponse> => [status, await query, headers]