+++ packages/axios-mock-server/src/lib/getInputs.ts
@@ -1,4 +1,0 @@
-import { defaultConfig } from './getConfig'
-
-export default (input?: string | string[]) =>
-  Array.isArray(input) ? input : [input || defaultConfig.input]