+++ packages/axios-mock-server/src/lib/writeRouteFile.ts
@@ -1,6 +1,0 @@
-import fs from 'fs'
-
-export default ({ filePath, text }: { filePath: string; text: string }) => {
-  fs.writeFileSync(filePath, text, 'utf8')
-  console.log(`${filePath} was built successfully.`)
-}